refactor: :recycle: implement proper seperation of concerns
