+++ vue-component/package.json
@@ -2,10 +2,10 @@
   "name": "crossfade-image-viewer",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/crossfade-image-viewer.ssr.js",
-  "browser": "dist/crossfade-image-viewer.esm.js",
-  "module": "dist/crossfade-image-viewer.esm.js",
-  "unpkg": "dist/crossfade-image-viewer.min.js",
+  "main": "dist/crossfade-image-viewer/crossfade-image-viewer.ssr.js",
+  "browser": "dist/crossfade-image-viewer/crossfade-image-viewer.esm.js",
+  "module": "dist/crossfade-image-viewer/crossfade-image-viewer.esm.js",
+  "unpkg": "dist/crossfade-image-viewer/crossfade-image-viewer.min.js",
   "types": "dist/types/src/entry.esm.d.ts",
   "files": [
     "dist/*",
